Refactor - Frontend
